+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/services/util/RequestInputUtilsTest.java
@@ -15,12 +15,13 @@ import org.junit.Test;
 import org.opendaylight.openflowplugin.api.OFConstants;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.MultipartType;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.MultipartRequestInput;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 public class RequestInputUtilsTest {
     @Test
     public void createMultipartHeader() {
         final Short version = OFConstants.OFP_VERSION_1_3;
-        final Long xid = 42L;
+        final Uint32 xid = Uint32.valueOf(42L);
         final MultipartType type = MultipartType.OFPMPDESC;
 
         final MultipartRequestInput input = RequestInputUtils
